Removed unnecessary 'sidebar' class & renamed SCSS
The 'sidebar' CSS selector for <aside> was unnecessary and confusing, since the <nav> element was an already defined container for the navigation sidebar. So, it was changed so that <aside> just acted as an indicator for screen readers that the navigation bar is placed on the left side. Also, an underscore ('_') was added before SCSS partial files to align with standards