Added JS code for an SPA, updated navigation bar
Added some of the JS code for an SPA by following a tutorial. Updated the navigation bar items by increasing the size of the font, reducing font-weight. Removed 'Add a Recipe' and 'Add a Recipe Log' as they were changed to modal forms. Added 'My Baking Logs' and renamed 'Dessert Recipes' to 'My Recipes' for clarity. Aligned the hamburger menu icon with the logo in the header with SCSS.